+++ lua/plugins/config/surround_config.lua
@@ -6,14 +6,24 @@ return function(use)
 			local prequire = require('personal.utils.prequire')
 			local surround = require('surround')
 			surround.setup({
-				prefix = 'ss',
+				prefix = 'q',
 			})
 			local wk = prequire('which-key')
 			if not wk then
 				return
 			end
+
 			wk.register({
-				ss = { 'Surround Prefix' },
+				qa = 'Add Surround <Motion>',
+				qq = 'Repeat Last Surround',
+				qd = 'Delete Surround <Motion>',
+				qr = 'Replace Surround <Motion>',
+				qt = {
+					name = 'toggle brackets',
+					b = 'Toggle Brackets',
+					q = 'Toggle Quotes',
+					B = 'Toggle Brackets',
+				},
 			})
 		end,
 	})